+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Log more meaningful messages in warnings (#268)
+
 ## [3.1.0] - 2025-07-22
 
 ### Changed