+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (as of version 2.0.11).
 
-## [Unreleased]
+## [3.2.0] - 2026-09-07
 
 ### Changed
 
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Changed the `data.js` to `data_{lang}.js` and `data_{lang}_{slug}.js` to reduce file size (#257)
 - Log more meaningful messages in warnings (#268)
 - Better warnings on missing language names (#269)
 - Do not crash the whole scrape with a `KeyError: 'metadata'` when a talk's `playerData.resources` has no `hls.metadata` entry (#261)
@@ -26,7 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Changed the `data.js` to `data_{lang}.js` and `data_{lang}_{slug}.js` to reduce file size. (#257)
 - Add `--language-threshold` CLI argument for considering languages that appear in at least specified percentage of videos in `compute_zim_languages` (#212)
 - Circular back-to-top button and replaced text characters with icon buttons for better visual consistency (#112)
 - Allow users to only get specific video from TED instead extract the entire topic or playlist (#237)